[Eclipse Foundation] Re-license the repo
This repo is to be moved to the Eclipse Theia project. First we
have to adopt the same license:

EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0

This commit adds this license and copyright headers, following
the re-licensing exercise. See here for the details:
https://github.com/theia-ide/theia-website/wiki/Re-licensing-theia-ide-theia-website

Note: originally there were no copyright headers in place, in this
repo. So instead of updating existhing headers to reflect the new
license, I created new headers, giving credit to the original IP
rights holder, to the best of my after-the-fact git detective abilities.
At the end of the day, the git history contains the details of
where every character in the repo came from, and that's of course the
source of thruth, not the results of my good faith attempt to figure-out
who contributed each source code file, here.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>