Prevent layout shifting of landing page
Previously the list of Contributors & Adopters was queried from the Eclipse Foundation
API after the initial rendering of the landing page. Once they were loaded they were
inserted leading to a layout shift. This was especially noticeable when visting the page
via an anchor as the content moved out of the frame.

This is now fixed by determining the list of Contributors & Adopters at build time.
As they are no longer loaded dynamically the layout shift is avoided. Additionally we no
longer need to wait for the API call to finish when visting the website, improving the
load times.

The website needs to be rebuilt whenever the Contributors & Adopters list shall be
updated. Should the API call fail the website build will panic to avoid publishing the
website with an incomplete list.

Contributed on behalf of STMicroelectronics