Add workflows for publishing via GH pages
This change adds a Github workflow that builds the website and
publishes the build result to branch `gh-pages` for every commit
on `master`. In addition, it deploys a preview to
https://github.com/eclipse-theia/theia-website-previews.

Effectively, this is the first step of changing the hosting of the
Theia website from Netlify to Github pages.

For more information see
https://gitlab.eclipse.org/eclipsefdn/helpdesk/-/issues/3874

This change applies the same technique as used in
https://github.com/eclipse-langium/langium-website

Co-authored-by: Olaf Lessenich <olessenich@eclipsesource.com>