Add automated publishing to `gh-pages`
This change adds a Github workflow that builds the website and publishes
the build result to branch `gh-pages`.

Thus, effectively, this is the first step of changing the hosting of the
Theia website from Netlify to Github pages.

Therefore, we also requested a new Github secret for publishing
and eventually will request changing the target of theia-ide.org.
For more information see
https://gitlab.eclipse.org/eclipsefdn/helpdesk/-/issues/3874