Fix image paths so they work with path prefix
Referencing images with a URL starting with `/` will always default to the root. This however won't work if we host the page in a folder. As we use plain HTML tags in the markdown, probably the easiest fix is to use relative image links instead.

This change turns all root paths for images into relative paths.