Fix image paths and links so they work with path prefix (#540)
This change turns all root paths for images and links into relative paths so they work also when we run the web page under a specific sub-folder below a domain.

Referencing images or other pages with a URL starting with `/` will always default to the root folder of a domain. This however won't work if we host the page in a sub-folder of a domain. As we use plain HTML tags in the markdown, probably the easiest fix is to use relative image links instead.

Co-authored-by: Olaf Lessenich <olessenich@eclipsesource.com>